% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSignal.R
\name{countSignal}
\alias{countSignal}
\title{Count the Sites Carrying an Arbitrary Signal Inside Genomic Regions}
\usage{
countSignal(signal, gr, maxDist = NULL, ignore.strand = TRUE, verbose = FALSE)
}
\arguments{
\item{signal}{A GRange object carrying the sites of interest.}

\item{gr}{A GRange object carrying the regions of interest.}

\item{ignore.strand}{When set to TRUE, the strand information is ignored in
the overlap calculations.}

\item{verbose}{Logical. Default is TRUE. If TRUE, then the progress of the
computational tasks is given.}
}
\value{
A GRanges object with the genomic regions that carry signals and the
number of sites on them.
}
\description{
A simple function to count the number of sites carrying an 
arbitrary signal, which are inside given genomic regions.
}
\details{
Given a GRanges object 'signal' carrying an arbitrary signal on each 
range, this function counts the number of sites inside the the given genomic 
regions 'gr'.
}
\examples{
some.signal <- makeGRangesFromDataFrame(data.frame(chr = "chr1", 
                                                   start = 1:15,
                                                   end = 1:15,
                                                   strand = '*'))

some.regions <- makeGRangesFromDataFrame(data.frame(chr = "chr1", 
                                                    start = c(2, 8),
                                                    end = c(7, 14),
                                                    strand = '*'))

countSignal(signal = some.signal, gr = some.regions)

}
\seealso{
\code{\link{getGRegionsStat}}
}
\author{
Robersy Sanchez. \url{https://genomaths.com}
}
