% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_GRMatrix.R
\docType{data}
\name{GRMatrix}
\alias{GRMatrix}
\title{Simulated matrix of arbitrary methylation signal used in the examples}
\format{
GRMatrix is a GRanges object with six column in its metacolumns
named "CT1", "CT2", "CT3", "TT1", "TT2", and "TT3".
}
\usage{
GRMatrix
}
\description{
This is a GRanges object (named 'GRMatrix') carrying a simulated matrix of 
arbitrary methylation signal in the metacolumn.
}
\keyword{datasets}
