% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GammaMixt.R
\name{predict.GammaMixt}
\alias{predict.GammaMixt}
\title{Predict function for the DMP's Mixtures of Gamma Distributions model}
\usage{
\method{predict}{GammaMixt}(gmd, LR, pred = "quant", q = 0.95,
  div.col = NULL)
}
\arguments{
\item{gmd}{An object carrying the best nonlinear fit for a distribution model
obtained with function \code{\link{nonlinearFitDist}} ('GammaMixt' 
class).}

\item{LR}{Data used to generate the model. A "pDMP"or "InfDiv" object 
obtained with functions \code{\link[MethylIT]{getPotentialDIMP}} or
\code{\link[MethylIT]{estimateDivergence}}. These are list of GRanges
objects, where each GRanges object from the list must have at least two
columns: a column containing the total variation of methylation level
(TV, difference of methylation levels) and a column containing a
divergence of methylation levels (it could be TV or  Hellinger
divergence).}

\item{pred}{Type of prediction resquested: *density* ("dens"),*quantiles*
("quant"), *random number* ("rnum"), *probabilities* ("prob"), or
classification *posterior probability* ("postPrb").}

\item{q}{numeric vector of quantiles, probabilities or an interger if
pred = "rnum".}

\item{dist.name}{name of the distribution to fit: Weibull2P (default:
"Weibull2P"), Weibull three-parameters (Weibull3P), gamma with
three-parameter (Gamma3P), gamma with two-parameter (Gamma2P),
generalized gamma with three-parameter ("GGamma3P") or four-parameter
("GGamma4P").}

\item{num.cores, tasks}{Paramaters for parallele computation using package
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to
use, i.e. at most how many child processes will be run simultaneously
(see \code{\link[BiocParallel]{bplapply}} and the number of tasks per job
(only for Linux OS).}
}
\description{
This is an utility function to get posterior probability 
    predictions based on a given DMP's Mixtures of Gamma Distributions (GMD)
    model, obtained with functio \code{\link{gammaMixtCut}}.
}
\details{
Predictions are based on the best model fit returned by function
    \code{\link{nonlinearFitDist}}. The possible prediction are: *density*,
    *quantiles*, *random number* or *probabilities*.
}
\examples{

}
