% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannonEntr.R
\name{shannonEntr}
\alias{shannonEntr}
\title{Compute Shannon Entropy}
\usage{
shannonEntr(p, logbase = 2)
}
\arguments{
\item{p}{A probability vector, sum(p) = 1.}

\item{logbase}{A positive number: the base with respect to which logarithms}
}
\description{
Compute Shannon Entropy of probability vector p.
}
\details{
By definition, if p_i = 0 for some i, then the entropy is 0. In 
    addition, since Shannon entropy is only defined for \eqn{0 <= p <= 1},
    the function will return 0 for any values of p out of the definition
    range of values.
}
\examples{
counts = sample.int(10)
prob = counts/sum(counts)
shannonEntr(prob)

## Out of the definition range of values for p, 'shannonEntr' will return 0
shannonEntr(c(0.5, 1.2, -0.3))
}
\author{
Robersy Sanchez (\url{https://github.com/genomaths}).
}
