% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GammaMixt.R
\name{predict.GammaMixt}
\alias{predict.GammaMixt}
\title{Predict function for the DMP's Mixtures of Gamma Distributions model}
\usage{
\method{predict}{GammaMixt}(gmd, pred = "quant", q = 0.95, div.col = NULL, interval = NULL)
}
\arguments{
\item{gmd}{An object carrying the best nonlinear fit for a distribution model
obtained with function \code{\link{nonlinearFitDist}} ('GammaMixt' 
class).}

\item{pred}{Type of prediction resquested: *density* ('dens'),*quantiles*
('quant'), *random number* ('rnum'), *probabilities* ('prob'), or
classification *posterior probability* ('postPrb').}

\item{q}{numeric vector of quantiles, probabilities or an interger if 
pred = 'rnum', or A 'pDMP'or 'InfDiv' object obtained with functions
\code{\link[MethylIT]{getPotentialDIMP}} or
\code{\link[MethylIT]{estimateDivergence}}. These are list of GRanges
objects, where each GRanges object from the list must have at least two
columns: a column containing the total variation of methylation level
(TV, difference of methylation levels) and a column containing a
divergence of methylation levels (it could be TV or  Hellinger
divergence).}

\item{div.col}{An integer. If 'q' is 'pDMP'or 'InfDiv' object, then it is the 
column number for the divergence of methylation levels used in the
estimation of model 'gmd'. Default: NULL.}

\item{interval}{a vector containing the end-points of the interval to be 
searched for the quantile(s). An interval would be, e.g., 'interval =
c(min(x), max(x))', where 'x' is the variable used to estimate model
'gmd'.}

\item{num.cores, tasks}{Paramaters for parallele computation using package
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to
use, i.e. at most how many child processes will be run simultaneously
(see \code{\link[BiocParallel]{bplapply}} and the number of tasks per job
(only for Linux OS).}
}
\description{
This is an utility function to get the density, probailities,
    and posterior probability predictions based on a given DMP's Mixtures of
    Gamma Distributions (GMD) model, obtained with function
    \code{\link{gammaMixtCut}}.
}
\details{
Predictions are based on the best model fit returned by function
    \code{\link{nonlinearFitDist}}. The possible prediction are: *density*,
    *quantiles*, *random number* or *probabilities*.
}
\examples{

}
