% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCounts.R
\name{simulateCounts}
\alias{simulateCounts}
\title{Simulate read counts of methylated and unmethylated cytosines}
\usage{
simulateCounts(num.samples, sites, alpha, beta, size, theta,
  sample.ids = NULL, chromosome = "1", start = NULL, end = NULL,
  strand = "*")
}
\arguments{
\item{num.samples}{Number of samples to generate.}

\item{sites}{Number of cytosine sites for each sample.}

\item{alpha}{Alpha parameter of beta distribution. Parameter shape1 from
\code{\link[stats]{Beta}} function.}

\item{beta}{Beta parameter of beta distribution. Parameter shape2 from
\code{\link[stats]{Beta}} function.}

\item{size}{number of trials (11 or more). Expected cytosine coverage.}

\item{theta}{Parameter theta from \code{\link[MASS]{rnegbin}}
(overdispersion parameter).}

\item{sample.ids}{Names for the samples.}

\item{chromosome}{A character string naming the chromosome. Default "1".}

\item{start}{An nteger vector with the start position for each cytosine site.
Default start = seq_len(sites).}

\item{end}{An integer vector with the end position for each cytosine site.
Default end = start. Notice that if end > start, each counts will cover
a region.}

\item{strand}{One of the characters '*', '+', or '-' denoting the DNA strand.
Default is '*'.}
}
\value{
A list of GRanges objects with the methylated and unmethylated counts
    in its metacolumn.
}
\description{
Auxiliary function to simulate read counts of methylated and
    unmethylated cytosines
}
\details{
Methylation coverages (minimum 10) are generated from a Negative
    Binomial distribution with function \code{\link[MASS]{rnegbin}} from R
    package MASS. This function uses the representation of the Negative
    Binomial distribution as a continuous mixture of Poisson distributions
    with Gamma distributed means. Prior methylation levels are randomly
    generated with beta distribution using \code{\link[stats]{Beta}}
    function from R package “stats” and posterior methylation levels are
    generated according Bayes' theorem. The read of methylation counts are
    obtained as the product of coverage by the posterior methylation level.
}
\examples{
# *** Simulate samples with expected average of difference of methylation
# levels equal to 0.0427.
# === Expected mean of methylation levels ===
bmean <- function(alpha, beta) alpha/(alpha + beta)
bmean(0.03, 0.5) - bmean(0.007, 0.5) #' Expected difference = 0.04279707

# === The number of cytosine sitesto generate ===
sites = 5000
# == Set a seed for pseudo-random number generation ===
set.seed(123)

# === Simulate samples ===
ref = simulateCounts(num.samples = 1, sites = sites, alpha = 0.007,
                    beta = 0.5, size = 50, theta = 4.5, sample.ids = "C1")
treat = simulateCounts(num.samples = 2, sites = sites, alpha = 0.03,
                    beta = 0.5, size = 50, theta = 4.5,
                    sample.ids = c("T1", "T2"))
}
\author{
Robersy Sanchez
}
